working version.
Now builds and works. Threading comes for free, no recompile needed.
Plan 9 support is split out.

+++ ftqcore.c
@@ -29,7 +29,7 @@ void           *
 ftq_core(void *arg)
 {
 	/* thread number, zero based. */
-	int             thread_num = (unsigned int) arg;
+	int             thread_num = (uintptr_t) arg;
 	int             i, offset;
 	int             k;
 	ticks           now, last, endinterval;
